Backend/requirements.txt
asgiref==3.2.10
certifi==2020.12.5
dj-database-url==0.5.0
Django
djangorestframework 
gunicorn==20.0.4
PyJWT 
pytz==2020.1
sqlparse 
whitenoise==5.2.0
